Explicit cast to (uint8_t)
In order for
    slen = (*np)->slen - mlen;
to be safely[1] cast to (uint8_t)slen, we need
    (*np)->slen >= mlen
which comes down to the behaviour of the
    compare(*np, key, keylen, &mlen, &mask)
function, which states
    "set the values of mlen... to the number of matching bytes"

Checking the source code for compare() confirms that mlen will indeed be
<= slen.  Therefore the cast is safe.

[1] yes, an underflow in unsigned integers is not undefined behaviour, but I'm
happier if we can avoid it.

+++ lib/datastruct/patricia.c
@@ -340,7 +340,7 @@ patricia_insert(PATRICIA * P, const uint8_t * key, size_t keylen, void * rec)
 			 */
 			/* Create the lower of the new nodes. */
 			slen = (*np)->slen - mlen;
-			pnew2 = node_dup(*np, slen, (*np)->s + mlen);
+			pnew2 = node_dup(*np, (uint8_t)slen, (*np)->s + mlen);
 			if (pnew2 == NULL)
 				return (-1);
